+++ internal/pkg/plan/plan.go
@@ -19,7 +19,6 @@ package plan
 import (
 	"fmt"
 	"github.com/pkg/errors"
-	"github.com/sirupsen/logrus"
 	"github.com/sugarkube/sugarkube/internal/pkg/cmd/cli/cluster"
 	"github.com/sugarkube/sugarkube/internal/pkg/constants"
 	"github.com/sugarkube/sugarkube/internal/pkg/installer"
@@ -98,7 +97,7 @@ func Create(forward bool, stackObj interfaces.IStack, manifests []interfaces.IMa
 	var previousManifest interfaces.IManifest
 
 	for _, installableObj := range selectedInstallables {
-		var installDestroyTask *task
+		var installDeleteTask *task
 
 		// create a new tranche for each new manifest
 		if previousManifest != nil && previousManifest.Id() != installableObj.ManifestId() &&
@@ -115,22 +114,22 @@ func Create(forward bool, stackObj interfaces.IStack, manifests []interfaces.IMa
 
 		// when creating a forward plan, we can install kapps...
 		if forward && installableObj.State() == constants.PresentKey {
-			installDestroyTask = &task{
+			installDeleteTask = &task{
 				installableObj: installableObj,
 				action:         constants.TaskActionInstall,
 			}
-			// but reverse plans must always destroy them
+			// but reverse plans must always delete them
 		} else if !forward || installableObj.State() == constants.AbsentKey {
-			installDestroyTask = &task{
+			installDeleteTask = &task{
 				installableObj: installableObj,
-				action:         constants.TaskActionDestroy,
+				action:         constants.TaskActionDelete,
 			}
 		}
 
-		if installDestroyTask != nil {
-			log.Logger.Debugf("Adding %s task for kapp '%s'", installDestroyTask.action,
+		if installDeleteTask != nil {
+			log.Logger.Debugf("Adding %s task for kapp '%s'", installDeleteTask.action,
 				installableObj.FullyQualifiedId())
-			tasks = append(tasks, *installDestroyTask)
+			tasks = append(tasks, *installDeleteTask)
 		}
 
 		// when tearing down a cluster users may not want to execute any post-actions
@@ -248,7 +247,7 @@ func reverse(plan *Plan) {
 }
 
 // Run a plan to make a target cluster have the necessary kapps installed/
-// destroyed to match the input manifests. Each tranche is run sequentially,
+// deleted to match the input manifests. Each tranche is run sequentially,
 // and each kapp in each tranche is processed in parallel.
 func (p *Plan) Run(approved bool, dryRun bool) error {
 
@@ -298,13 +297,6 @@ func (p *Plan) Run(approved bool, dryRun bool) error {
 		for success := 0; success < totalOperations; success++ {
 			select {
 			case err := <-errCh:
-				if log.Logger.Level == logrus.DebugLevel {
-					log.Logger.Fatalf("Error processing kapp in tranche %d of plan: %+v", i+1, err)
-				} else {
-					log.Logger.Fatalf("Error processing kapp in tranche %d of plan: %v\n"+
-						"Run with `-l debug` for a full stacktrace.", i+1, err)
-				}
-				close(doneCh)
 				return errors.Wrapf(err, "Error processing kapp goroutine "+
 					"in tranche %d of plan", i+1)
 			case <-doneCh:
@@ -319,7 +311,7 @@ func (p *Plan) Run(approved bool, dryRun bool) error {
 	return nil
 }
 
-// Installs or destroys a kapp using the appropriate Installer
+// Installs or deletes a kapp using the appropriate Installer
 func processKapp(jobs <-chan job, doneCh chan bool, errCh chan error) {
 
 	for job := range jobs {
@@ -351,15 +343,15 @@ func processKapp(jobs <-chan job, doneCh chan bool, errCh chan error) {
 
 		switch task.action {
 		case constants.TaskActionInstall:
-			err := installer.Install(installerImpl, installableObj, stackObj, approved, renderTemplates, dryRun)
+			err := installerImpl.Install(installableObj, stackObj, approved, renderTemplates, dryRun)
 			if err != nil {
 				errCh <- errors.Wrapf(err, "Error installing kapp '%s'", installableObj.Id())
 			}
 			break
-		case constants.TaskActionDestroy:
-			err := installer.Destroy(installerImpl, installableObj, stackObj, approved, renderTemplates, dryRun)
+		case constants.TaskActionDelete:
+			err := installerImpl.Delete(installableObj, stackObj, approved, renderTemplates, dryRun)
 			if err != nil {
-				errCh <- errors.Wrapf(err, "Error destroying kapp '%s'", installableObj.Id())
+				errCh <- errors.Wrapf(err, "Error deleting kapp '%s'", installableObj.Id())
 			}
 			break
 		case constants.TaskActionClusterUpdate: